refactor: fusionner components/analysis dans components/analyse-video (analysis-modal etait sous-composant de video-task-editor)
